requirements.txt
Flask>=2.3.0
PyYAML>=6.0
PyMuPDF>=1.23.0
openai>=1.3.5
Werkzeug>=2.3.0
